feat(Tasks): implement TasksPage component with task management functionality
